Add new conftest file in _files/
