+++ tests/_files/test_inputmedia.py
@@ -36,31 +36,13 @@ from telegram import (
     ReplyParameters,
 )
 from telegram.constants import InputMediaType, ParseMode
-
-# noinspection PyUnresolvedReferences
 from telegram.error import BadRequest
 from telegram.request import RequestData
-from tests._files.test_animation import animation, animation_file  # noqa: F401
 from tests.auxil.files import data_file
 from tests.auxil.networking import expect_bad_request
 from tests.auxil.slots import mro_slots
 
-# noinspection PyUnresolvedReferences
-from tests.test_forum import emoji_id, real_topic  # noqa: F401
-
 from ..auxil.build_messages import make_message
-
-# noinspection PyUnresolvedReferences
-from .test_audio import audio, audio_file  # noqa: F401
-
-# noinspection PyUnresolvedReferences
-from .test_document import document, document_file  # noqa: F401
-
-# noinspection PyUnresolvedReferences
-from .test_photo import photo, photo_file, photolist, thumb  # noqa: F401
-
-# noinspection PyUnresolvedReferences
-from .test_video import video, video_file  # noqa: F401
 
 
 @pytest.fixture(scope="module")
@@ -183,7 +165,7 @@ class TestInputMediaVideoWithoutRequest(TestInputMediaVideoBase):
         assert input_media_video_dict["supports_streaming"] == input_media_video.supports_streaming
         assert input_media_video_dict["has_spoiler"] == input_media_video.has_spoiler
 
-    def test_with_video(self, video):  # noqa: F811
+    def test_with_video(self, video):
         # fixture found in test_video
         input_media_video = InputMediaVideo(video, caption="test 3")
         assert input_media_video.type == self.type_
@@ -193,7 +175,7 @@ class TestInputMediaVideoWithoutRequest(TestInputMediaVideoBase):
         assert input_media_video.duration == video.duration
         assert input_media_video.caption == "test 3"
 
-    def test_with_video_file(self, video_file):  # noqa: F811
+    def test_with_video_file(self, video_file):
         # fixture found in test_video
         input_media_video = InputMediaVideo(video_file, caption="test 3")
         assert input_media_video.type == self.type_
@@ -267,14 +249,14 @@ class TestInputMediaPhotoWithoutRequest(TestInputMediaPhotoBase):
         ]
         assert input_media_photo_dict["has_spoiler"] == input_media_photo.has_spoiler
 
-    def test_with_photo(self, photo):  # noqa: F811
+    def test_with_photo(self, photo):
         # fixture found in test_photo
         input_media_photo = InputMediaPhoto(photo, caption="test 2")
         assert input_media_photo.type == self.type_
         assert input_media_photo.media == photo.file_id
         assert input_media_photo.caption == "test 2"
 
-    def test_with_photo_file(self, photo_file):  # noqa: F811
+    def test_with_photo_file(self, photo_file):
         # fixture found in test_photo
         input_media_photo = InputMediaPhoto(photo_file, caption="test 2")
         assert input_media_photo.type == self.type_
@@ -332,14 +314,14 @@ class TestInputMediaAnimationWithoutRequest(TestInputMediaAnimationBase):
         assert input_media_animation_dict["duration"] == input_media_animation.duration
         assert input_media_animation_dict["has_spoiler"] == input_media_animation.has_spoiler
 
-    def test_with_animation(self, animation):  # noqa: F811
+    def test_with_animation(self, animation):
         # fixture found in test_animation
         input_media_animation = InputMediaAnimation(animation, caption="test 2")
         assert input_media_animation.type == self.type_
         assert input_media_animation.media == animation.file_id
         assert input_media_animation.caption == "test 2"
 
-    def test_with_animation_file(self, animation_file):  # noqa: F811
+    def test_with_animation_file(self, animation_file):
         # fixture found in test_animation
         input_media_animation = InputMediaAnimation(animation_file, caption="test 2")
         assert input_media_animation.type == self.type_
@@ -400,7 +382,7 @@ class TestInputMediaAudioWithoutRequest(TestInputMediaAudioBase):
             ce.to_dict() for ce in input_media_audio.caption_entities
         ]
 
-    def test_with_audio(self, audio):  # noqa: F811
+    def test_with_audio(self, audio):
         # fixture found in test_audio
         input_media_audio = InputMediaAudio(audio, caption="test 3")
         assert input_media_audio.type == self.type_
@@ -410,7 +392,7 @@ class TestInputMediaAudioWithoutRequest(TestInputMediaAudioBase):
         assert input_media_audio.title == audio.title
         assert input_media_audio.caption == "test 3"
 
-    def test_with_audio_file(self, audio_file):  # noqa: F811
+    def test_with_audio_file(self, audio_file):
         # fixture found in test_audio
         input_media_audio = InputMediaAudio(audio_file, caption="test 3")
         assert input_media_audio.type == self.type_
@@ -471,14 +453,14 @@ class TestInputMediaDocumentWithoutRequest(TestInputMediaDocumentBase):
             == input_media_document.disable_content_type_detection
         )
 
-    def test_with_document(self, document):  # noqa: F811
+    def test_with_document(self, document):
         # fixture found in test_document
         input_media_document = InputMediaDocument(document, caption="test 3")
         assert input_media_document.type == self.type_
         assert input_media_document.media == document.file_id
         assert input_media_document.caption == "test 3"
 
-    def test_with_document_file(self, document_file):  # noqa: F811
+    def test_with_document_file(self, document_file):
         # fixture found in test_document
         input_media_document = InputMediaDocument(document_file, caption="test 3")
         assert input_media_document.type == self.type_
@@ -494,7 +476,7 @@ class TestInputMediaDocumentWithoutRequest(TestInputMediaDocumentBase):
 
 
 @pytest.fixture(scope="module")
-def media_group(photo, thumb):  # noqa: F811
+def media_group(photo, thumb):
     return [
         InputMediaPhoto(photo, caption="*photo* 1", parse_mode="Markdown"),
         InputMediaPhoto(thumb, caption="<b>photo</b> 2", parse_mode="HTML"),
@@ -505,12 +487,12 @@ def media_group(photo, thumb):  # noqa: F811
 
 
 @pytest.fixture(scope="module")
-def media_group_no_caption_args(photo, thumb):  # noqa: F811
+def media_group_no_caption_args(photo, thumb):
     return [InputMediaPhoto(photo), InputMediaPhoto(thumb), InputMediaPhoto(photo)]
 
 
 @pytest.fixture(scope="module")
-def media_group_no_caption_only_caption_entities(photo, thumb):  # noqa: F811
+def media_group_no_caption_only_caption_entities(photo, thumb):
     return [
         InputMediaPhoto(photo, caption_entities=[MessageEntity(MessageEntity.BOLD, 0, 5)]),
         InputMediaPhoto(photo, caption_entities=[MessageEntity(MessageEntity.BOLD, 0, 5)]),
@@ -518,7 +500,7 @@ def media_group_no_caption_only_caption_entities(photo, thumb):  # noqa: F811
 
 
 @pytest.fixture(scope="module")
-def media_group_no_caption_only_parse_mode(photo, thumb):  # noqa: F811
+def media_group_no_caption_only_parse_mode(photo, thumb):
     return [
         InputMediaPhoto(photo, parse_mode="Markdown"),
         InputMediaPhoto(thumb, parse_mode="HTML"),
@@ -549,10 +531,10 @@ class TestSendMediaGroupWithoutRequest:
         self,
         bot,
         chat_id,
-        photo_file,  # noqa: F811
-        animation_file,  # noqa: F811
-        audio_file,  # noqa: F811
-        video_file,  # noqa: F811
+        photo_file,
+        animation_file,
+        audio_file,
+        video_file,
         monkeypatch,
     ):
         async def make_assertion(url, request_data: RequestData, *args, **kwargs):
@@ -576,7 +558,7 @@ class TestSendMediaGroupWithoutRequest:
             await bot.send_media_group(chat_id, media)
 
     async def test_send_media_group_with_thumbs(
-        self, bot, chat_id, video_file, photo_file, monkeypatch  # noqa: F811
+        self, bot, chat_id, video_file, photo_file, monkeypatch
     ):
         async def make_assertion(method, url, request_data: RequestData, *args, **kwargs):
             nonlocal input_video
@@ -594,7 +576,7 @@ class TestSendMediaGroupWithoutRequest:
             await bot.send_media_group(chat_id, [input_video, input_video])
 
     async def test_edit_message_media_with_thumb(
-        self, bot, chat_id, video_file, photo_file, monkeypatch  # noqa: F811
+        self, bot, chat_id, video_file, photo_file, monkeypatch
     ):
         async def make_assertion(
             method: str, url: str, request_data: Optional[RequestData] = None, *args, **kwargs
@@ -663,9 +645,7 @@ class TestSendMediaGroupWithRequest:
             mes.caption_entities == (MessageEntity(MessageEntity.BOLD, 0, 5),) for mes in messages
         )
 
-    async def test_send_media_group_new_files(
-        self, bot, chat_id, video_file, photo_file  # noqa: F811
-    ):
+    async def test_send_media_group_new_files(self, bot, chat_id, video_file, photo_file):
         async def func():
             return await bot.send_media_group(
                 chat_id,
@@ -701,7 +681,7 @@ class TestSendMediaGroupWithRequest:
         assert all(mes.media_group_id == messages[0].media_group_id for mes in messages)
 
     async def test_send_media_group_with_message_thread_id(
-        self, bot, real_topic, forum_group_id, media_group  # noqa: F811
+        self, bot, real_topic, forum_group_id, media_group
     ):
         messages = await bot.send_media_group(
             forum_group_id,
@@ -800,9 +780,7 @@ class TestSendMediaGroupWithRequest:
             )
             assert all(mes.has_protected_content for mes in messages)
 
-    async def test_send_media_group_with_spoiler(
-        self, bot, chat_id, photo_file, video_file  # noqa: F811
-    ):
+    async def test_send_media_group_with_spoiler(self, bot, chat_id, photo_file, video_file):
         # Media groups can't contain Animations, so that is tested in test_animation.py
         media = [
             InputMediaPhoto(photo_file, has_spoiler=True),
@@ -945,11 +923,11 @@ class TestSendMediaGroupWithRequest:
         chat_id,
         default_bot,
         media_type,
-        animation,  # noqa: F811
-        document,  # noqa: F811
-        audio,  # noqa: F811
-        photo,  # noqa: F811
-        video,  # noqa: F811
+        animation,
+        document,
+        audio,
+        photo,
+        video,
     ):
         html_caption = "<b>bold</b> <i>italic</i> <code>code</code>"
         markdown_caption = "*bold* _italic_ `code`"
